Align chili page with assignment
Updated the prchili.html page to better align it with the actual
assignment a la Odin Project. As a rule, I've always been a "I get the
idea" kind of guy and leap into it. While my intuition was close, it
wasn't spot on and I need to get into the habit of reading instructions
completely instead of jumping the gun.

For what it's worth, I'm still proud of what I've done. And thoroughly
enjoying this process as a whole.