Adjust prchili link
Removed target=blank and rel=noopener noreferrer from the chili link,
as from a security perspective it is self-contained in the own site.
Not sure that it's ultimately relevent, but it seems to be a logical
change and functionality that can be followed with future internal links
in an attempt to save some keystrokes in development.